Chore: Refactor the snapshot state from the state sync
